Added variable names to the model. More information in the Description.
There is an issue where Gurobi thinks that there are twice as many π variables as in the JuMP model. This should help with debugging and finding the cause. Now we can inspect the model at any point by simply 'println(model)' and the previously anonymous variables now have proper names. This helps new contributors tremendously because they can see the model formulation. Feel free to change the index notation if you come up with something that looks nicer. Notice, however, that the indices of the variables should match the syntax that was used before but was not visible.